refactor: rename file to keep with conventions
