+++ packages/grafana-prometheus/src/components/monaco-query-field/MonacoQueryField.tsx
@@ -13,7 +13,7 @@ import { Monaco, monacoTypes, ReactMonacoEditor, useTheme2 } from '@grafana/ui';
 import { Props } from './MonacoQueryFieldProps';
 import { getOverrideServices } from './getOverrideServices';
 import { getCompletionProvider, getSuggestOptions } from './monaco-completion-provider';
-import { DataProvider } from './monaco-completion-provider/data-provider';
+import { DataProvider } from './monaco-completion-provider/data_provider';
 import { placeHolderScopedVars, validateQuery } from './monaco-completion-provider/validation';
 import { language, languageConfiguration } from './promql';
