+++ packages/grafana-prometheus/src/components/monaco-query-field/monaco-completion-provider/data_provider.ts
